% Auto-generated: do not edit by hand
\name{dashJXGBoard}

\alias{dashJXGBoard}

\title{JXGBoard component}

\description{
Displays Board of JSXGraph
}

\usage{
dashJXGBoard(id=NULL, boardAttributes=NULL, example=NULL, style=NULL,
value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{boardAttributes}{Named list. }

\item{example}{Numeric. 0 or 1}

\item{style}{Named list. style="width:500px; height:500px;"}

\item{value}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

